comments + cleanup
