+++ src/utils.ts
@@ -12,14 +12,18 @@ type EndpointFn = (req: Request, res: Response) => Promise<void>;
 
 /**
  * `throw` wrapped in a function, so we can use it in null coalescing statements.
- *
- * Don't ask me, ask whoever designed typescript this way.
  */
 export const throwFn = (e: Error): never => {
   throw e;
 };
 
+/**
+ * Wraps an asynchronous endpoint function with error handling.
+ * Any errors thrown by the endpoint function will be passed to the `next` function.
+ *
+ * @param fn - The asynchronous endpoint function to wrap.
+ * @returns A middleware function that handles errors thrown by the endpoint function.
+ */
 export const asyncCatch =
-  (fn: EndpointFn) => (req: Request, res: Response, next: NextFunction) => {
+  (fn: EndpointFn) => (req: Request, res: Response, next: NextFunction) =>
     fn(req, res).catch(next);
-  };
